+++ extruct/__init__.py
@@ -1,16 +1,7 @@
-from lxml.html import fromstring
-
-from extruct.jsonld import JsonLdExtractor
-from extruct.rdfa import RDFaExtractor
-from extruct.w3cmicrodata import MicrodataExtractor
-from extruct.xmldom import XmlDomHTMLParser
-
-
-def extract(htmlstring, url='http://www.example.com/', encoding="UTF-8"):
-    domparser = XmlDomHTMLParser(encoding=encoding)
-    tree = fromstring(htmlstring, parser=domparser)
-    return {name: extractor.extract_items(tree, url=url)
-            for name, extractor in (
-                ('json-ld', JsonLdExtractor()),
-                ('microdata', MicrodataExtractor()),
-                ('rdfa', RDFaExtractor()))}
+from ._extruct import SYNTAXES, extract
+from .jsonld import JsonLdExtractor
+from .rdfa import RDFaExtractor
+from .w3cmicrodata import MicrodataExtractor
+from .opengraph import OpenGraphExtractor
+from .microformat import MicroformatExtractor
+from .xmldom import XmlDomHTMLParser