+++ extruct/w3cmicrodata.py
@@ -10,13 +10,38 @@ follows http://www.w3.org/TR/microdata/#json
 """
 
 import collections
+from functools import partial
+
 try:
     from urlparse import urljoin
 except ImportError:
     from urllib.parse import urljoin
 
 import lxml.etree
-import lxml.html
+from lxml.html.clean import Cleaner
+from w3lib.html import strip_html5_whitespace
+import html_text
+
+from extruct.utils import parse_html
+
+
+# Cleaner which is similar to html_text cleaner, but is less aggressive
+cleaner = Cleaner(
+    scripts=True,
+    javascript=False,  # onclick attributes are fine
+    comments=True,
+    style=True,
+    links=True,
+    meta=True,
+    page_structure=False,  # <title> may be nice to have
+    processing_instructions=True,
+    embedded=False,  # keep embedded content
+    frames=False,  # keep frames
+    forms=False,  # keep forms
+    annoying_tags=False,
+    remove_unknown_tags=False,
+    safe_attrs_only=False,
+)
 
 
 class LxmlMicrodataExtractor(object):
@@ -31,34 +56,34 @@ class LxmlMicrodataExtractor(object):
                                        + count(ancestor::*[@itemscope])
                                        + 1""")
 
-    def __init__(self, nested=True, strict=False, add_text_content=False):
+    def __init__(self, nested=True, strict=False, add_text_content=False, add_html_node=False):
         self.nested = nested
         self.strict = strict
         self.add_text_content = add_text_content
+        self.add_html_node = add_html_node
 
     def get_docid(self, node):
         return int(self._xp_item_docid(node))
 
-    def extract(self, htmlstring, url='http://www.example.com/', encoding="UTF-8"):
-        parser = lxml.html.HTMLParser(encoding=encoding)
-        lxmldoc = lxml.html.fromstring(htmlstring, parser=parser)
-        return self.extract_items(lxmldoc, url)
+    def extract(self, htmlstring, base_url=None, encoding="UTF-8"):
+        tree = parse_html(htmlstring, encoding=encoding)
+        return self.extract_items(tree, base_url)
 
-    def extract_items(self, document, url, *args, **kwargs):
-        self.url = url
-        self.items_seen = set()
-        return [item
-                for item in map(self.extract_item,
-                                self._xp_item(document))
-                if item]
+    def extract_items(self, document, base_url):
+        items_seen = set()
+        return [
+            item for item in (
+                self._extract_item(it, items_seen=items_seen, base_url=base_url)
+                for it in self._xp_item(document))
+            if item]
 
-    def extract_item(self, node):
+    def _extract_item(self, node, items_seen, base_url):
         itemid = self.get_docid(node)
 
         if self.nested:
-            if itemid in self.items_seen:
+            if itemid in items_seen:
                 return
-            self.items_seen.add(itemid)
+            items_seen.add(itemid)
 
         item = {}
         if not self.nested:
@@ -75,15 +100,17 @@ class LxmlMicrodataExtractor(object):
                 item["id"] = itemid.strip()
 
         properties = collections.defaultdict(list)
-        # start with item references
+        for name, value in self._extract_properties(
+                node, items_seen=items_seen, base_url=base_url):
+            properties[name].append(value)
+
+        # process item references
         refs = node.get('itemref', '').split()
         if refs:
             for refid in refs:
-                for name, value in self.extract_property_refs(node, refid):
+                for name, value in self._extract_property_refs(
+                        node, refid, items_seen=items_seen, base_url=base_url):
                     properties[name].append(value)
-
-        for name, value in self.extract_properties(node):
-            properties[name].append(value)
 
         props = []
         for (name, values) in properties.items():
@@ -95,36 +122,58 @@ class LxmlMicrodataExtractor(object):
             item["properties"] = dict(props)
         else:
             # item without properties; let's use the node itself
-            item["value"] = self.extract_property_value(node, force=True)
+            item["value"] = self._extract_property_value(
+                node, force=True, items_seen=items_seen, base_url=base_url)
 
-        # not in the specs, but can be handy
+        # below are not in the specs, but can be handy
         if self.add_text_content:
-            textContent = self.extract_textContent(node)
+            textContent = self._extract_textContent(node)
             if textContent:
                 item["textContent"] = textContent
+        if self.add_html_node:
+            item["htmlNode"] = node
 
         return item
 
-    def extract_properties(self, node):
+    def _extract_properties(self, node, items_seen, base_url):
         for prop in self._xp_prop(node):
-            for p, v in self.extract_property(prop):
+            for p, v in self._extract_property(
+                    prop, items_seen=items_seen, base_url=base_url):
                 yield p, v
 
-    def extract_property_refs(self, node, refid):
-        for prop in node.xpath("id($refid)/descendant-or-self::*[@itemprop]", refid=refid):
-            for p, v in self.extract_property(prop):
+    def _extract_property_refs(self, node, refid, items_seen, base_url):
+        ref_node = node.xpath("id($refid)[1]", refid=refid)
+        if not ref_node:
+            return
+        ref_node = ref_node[0]
+        extract_fn = partial(self._extract_property, items_seen=items_seen,
+                             base_url=base_url)
+        if 'itemprop' in ref_node.keys() and 'itemscope' in ref_node.keys():
+            # An full item will be extracted from the node, no need to look
+            # for individual properties in childs
+            for p, v in extract_fn(ref_node):
                 yield p, v
+        else:
+            base_parent_scope = ref_node.xpath("ancestor-or-self::*[@itemscope][1]")
+            for prop in ref_node.xpath("descendant-or-self::*[@itemprop]"):
+                parent_scope = prop.xpath("ancestor::*[@itemscope][1]")
+                # Skip properties defined in a different scope than the ref_node
+                if parent_scope == base_parent_scope:
+                    for p, v in extract_fn(prop):
+                        yield p, v
 
-    def extract_property(self, node):
+    def _extract_property(self, node, items_seen, base_url):
         props = node.get("itemprop").split()
-        value = self.extract_property_value(node)
+        value = self._extract_property_value(
+            node, items_seen=items_seen, base_url=base_url)
         return [(p, value) for p in props]
 
-    def extract_property_value(self, node, force=False):
+    def _extract_property_value(self, node, items_seen, base_url, force=False):
         #http://www.w3.org/TR/microdata/#values
         if not force and node.get("itemscope") is not None:
             if self.nested:
-                return self.extract_item(node)
+                return self._extract_item(
+                    node, items_seen=items_seen, base_url=base_url)
             else:
                 return {"iid_ref": self.get_docid(node)}
 
@@ -132,13 +181,13 @@ class LxmlMicrodataExtractor(object):
             return node.get("content", "")
 
         elif node.tag in ("audio", "embed", "iframe", "img", "source", "track", "video"):
-            return urljoin(self.url, node.get("src", ""))
+            return urljoin(base_url, strip_html5_whitespace(node.get("src", "")))
 
         elif node.tag in ("a", "area", "link"):
-            return urljoin(self.url, node.get("href", ""))
+            return urljoin(base_url, strip_html5_whitespace(node.get("href", "")))
 
         elif node.tag in ("object",):
-            return urljoin(self.url, node.get("data", ""))
+            return urljoin(base_url, strip_html5_whitespace(node.get("data", "")))
 
         elif node.tag in ("data", "meter"):
             return node.get("value", "")
@@ -151,10 +200,11 @@ class LxmlMicrodataExtractor(object):
             return node.get("content")
 
         else:
-            return self.extract_textContent(node)
+            return self._extract_textContent(node)
 
-    def extract_textContent(self, node):
-        return u"".join(self._xp_clean_text(node)).strip()
+    def _extract_textContent(self, node):
+        clean_node = cleaner.clean_html(node)
+        return html_text.etree_to_text(clean_node)
 
 
 MicrodataExtractor = LxmlMicrodataExtractor