+++ tests/test_tool.py
@@ -6,9 +6,9 @@ try:
 except ImportError:
     import mock
 
-from extruct.tool import metadata_from_url
+from extruct.tool import metadata_from_url, main
 from requests.exceptions import HTTPError
-from tests import get_testdata
+from tests import get_testdata, jsonize_dict
 
 
 class TestTool(unittest.TestCase):
@@ -29,7 +29,7 @@ class TestTool(unittest.TestCase):
         mock_get.return_value = mock_response
 
         data = metadata_from_url(self.url)
-        self.assertEqual(data, expected)
+        self.assertEqual(jsonize_dict(data), expected)
 
     @mock.patch('extruct.tool.requests.get')
     def test_metadata_from_url_jsonld_only(self, mock_get):
@@ -44,8 +44,9 @@ class TestTool(unittest.TestCase):
         )
         mock_get.return_value = mock_response
 
-        data = metadata_from_url(self.url, microdata=False, jsonld=True, rdfa=False)
-        self.assertEqual(data, expected)
+        data = metadata_from_url(self.url, syntaxes=['json-ld'])
+        self.assertEqual(jsonize_dict(data), expected)
+
 
     @mock.patch('extruct.tool.requests.get')
     def test_metadata_from_url_microdata_only(self, mock_get):
@@ -60,8 +61,9 @@ class TestTool(unittest.TestCase):
         )
         mock_get.return_value = mock_response
 
-        data = metadata_from_url(self.url, microdata=True, jsonld=False, rdfa=False)
-        self.assertEqual(data, expected)
+        data = metadata_from_url(self.url, syntaxes=['microdata'])
+
+        self.assertEqual(jsonize_dict(data), expected)
 
     @mock.patch('extruct.tool.requests.get')
     def test_metadata_from_url_rdfa_only(self, mock_get):
@@ -76,8 +78,40 @@ class TestTool(unittest.TestCase):
         )
         mock_get.return_value = mock_response
 
-        data = metadata_from_url(self.url, microdata=False, jsonld=False, rdfa=True)
-        self.assertEqual(data, expected)
+        data = metadata_from_url(self.url, syntaxes=['rdfa'])
+        self.assertEqual(jsonize_dict(data), expected)
+
+    @mock.patch('extruct.tool.requests.get')
+    def test_metadata_from_url_opengraph_only(self, mock_get):
+        expected = {
+            'opengraph': self.expected['opengraph'],
+            'url': self.url,
+            'status': '200 OK',
+        }
+        mock_response = build_mock_response(
+            url=self.url,
+            content=get_testdata('songkick', 'tovestyrke.html'),
+        )
+        mock_get.return_value = mock_response
+
+        data = metadata_from_url(self.url, syntaxes=['opengraph'])
+        self.assertEqual(jsonize_dict(data), expected)
+
+    @mock.patch('extruct.tool.requests.get')
+    def test_metadata_from_url_microformat_only(self, mock_get):
+        expected = {
+            'microformat': self.expected['microformat'],
+            'url': self.url,
+            'status': '200 OK',
+        }
+        mock_response = build_mock_response(
+            url=self.url,
+            content=get_testdata('songkick', 'tovestyrke.html'),
+        )
+        mock_get.return_value = mock_response
+
+        data = metadata_from_url(self.url, syntaxes=['microformat'])
+        self.assertEqual(jsonize_dict(data), expected)
 
     @mock.patch('extruct.tool.requests.get')
     def test_metadata_from_url_unauthorized_page(self, mock_get):
@@ -95,6 +129,57 @@ class TestTool(unittest.TestCase):
         mock_response.raise_for_status.side_effect = http_error
 
         data = metadata_from_url(url)
+        self.assertEqual(data, expected)
+
+
+    @mock.patch('extruct.tool.requests.get')
+    def test_main_all(self, mock_get):
+        expected = self.expected
+        expected['url'] = self.url
+        expected['status'] = '200 OK'
+        expected = json.dumps(expected, indent=2, sort_keys=True)
+        mock_response = build_mock_response(
+            url=self.url,
+            content=get_testdata('songkick', 'tovestyrke.html'),
+        )
+        mock_get.return_value = mock_response
+
+        data = main([self.url])
+        self.assertEqual(data, expected)
+
+    @mock.patch('extruct.tool.requests.get')
+    def test_main_single_syntax(self, mock_get):
+        expected = {
+            'opengraph': self.expected['opengraph'],
+            'url': self.url,
+            'status': '200 OK',
+        }
+        expected = json.dumps(expected, indent=2, sort_keys=True)
+        mock_response = build_mock_response(
+            url=self.url,
+            content=get_testdata('songkick', 'tovestyrke.html'),
+        )
+        mock_get.return_value = mock_response
+
+        data = main([self.url, '--syntax', 'opengraph'])
+        self.assertEqual(data, expected)
+
+    @mock.patch('extruct.tool.requests.get')
+    def test_main_multiple_syntaxes(self, mock_get):
+        expected = {
+            'opengraph': self.expected['opengraph'],
+            'microdata': self.expected['microdata'],
+            'url': self.url,
+            'status': '200 OK',
+        }
+        expected = json.dumps(expected, indent=2, sort_keys=True)
+        mock_response = build_mock_response(
+            url=self.url,
+            content=get_testdata('songkick', 'tovestyrke.html'),
+        )
+        mock_get.return_value = mock_response
+
+        data = main([self.url, '--syntax', 'opengraph', 'microdata'])
         self.assertEqual(data, expected)